+++ application/api/controller/Menu.php
@@ -124,7 +124,10 @@ class Menu extends Base
     public function delete($id)
     {
         $menu = MenuModel::get($id);
-        
+        if (!$menu) {
+            return ajax_return(ResultCode::E_DATA_NOT_FOUND, '菜单不存在!');
+        }
+
         $res = $menu->delete();
         if (!$res) {
             return ajax_return(ResultCode::ACTION_FAILED, '操作失败!');
